requirements.txt
absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
executing==2.1.0
firebase-admin==6.5.0
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.6.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
h5py==3.12.1
httplib2==0.22.0
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.6.0
keyboard==0.13.5
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
msgpack==1.1.0
my-utils==0.1
-e git+https://github.com/Kanisorn-S/IOT_Project.git@d8989035c7e0ce9cd084fe6094d3708d597b1993#egg=mymodule&subdirectory=raspberrypi\utils
namex==0.0.8
nest-asyncio==1.6.0
numpy==2.0.2
opencv-contrib-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
paho-mqtt==2.1.0
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydot==3.0.2
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
pyzmq==26.2.0
requests==2.32.3
rich==13.9.4
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
-e git+https://github.com/Kanisorn-S/IOT_Project.git@d8989035c7e0ce9cd084fe6094d3708d597b1993#egg=utils&subdirectory=raspberrypi\utils
wcwidth==0.2.13
Werkzeug==3.1.1
wrapt==1.16.0
